requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
pandas==2.1.4
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
xlsxwriter==3.1.9 